Do not use mutable default arguments
Using mutable default arguments is a common Python problem, see e.g.
https://docs.python-guide.org/writing/gotchas/#mutable-default-arguments

In this specific case the default argument even tries to setup some
infrastructure settings at import time, which can potentially fail.

Note that this commit is changing the behavior for user passing in
`None` to the changed method, as it will now create an `AsyncClient`
with the options specified within the module. In the previous behavior
it would have raised an exception deeper down the call stack (see
http://github.com/microsoft/kiota-http-python/pull/383/files).